kra_jio_backend/requirements.txt
anyio==3.4.0
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Deprecated==1.2.13
et-xmlfile==1.1.0
fastapi==0.75.0
future==0.18.2
h11==0.12.0
idna==3.3
iniconfig==1.1.1
loguru==0.5.3
numpy==1.22.0
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pydantic==1.8.2
pyepc @ git+https://github.com/NiharZanwar/pyepc.git@e62b3dd5cd1519ec78c8285ac06a3b76a7b77623
pyfcm==1.5.4
PyJWT==1.7.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
redis==4.1.0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.17.6
wrapt==1.13.3